### 2021-05-23 #3307 @jacobmerson
Potential reviewers based on the joss_reviewers list starting at the bottom.
1. balos1
2. joeweaver
3. justingray (has modeling experience, C++ is secondary language which might be a challenge since I use a few more advanced features like SFINAE, forwarding refs, etc.)
4. atismer
5. zeroset
