### 2021-06-11 #3307 @joeweaver
I'll take a look at my calendar and the scope of the project and let you
know by the end of the day (UK time).

On Thu, Jun 10, 2021 at 9:49 PM Patrick Diehl ***@***.***>
wrote:

> Hey, @balos1 <https://github.com/balos1> @joeweaver
> <https://github.com/joeweaver> @justingray <https://github.com/justingray>
> @atismer <https://github.com/atismer> @zeroset
> <https://github.com/zeroset> would you be interested to review this paper?
>
