### 2021-06-11 #3307 @atismer
Generally yes, I'm interested to review the paper. But when reading the summary of the paper I was thinking if my limited experience in FEM is sufficient.
