### 2021-06-11 #3307 @joeweaver
This looks like a very useful project and I think it's really cool that the
other wrote it so that it has no hard dependencies on their simulation
tool, which appears to be Simmetrix.  However, I do not have Simmetrix
access nor sufficient time to commit to learning how apply the framework to
a different simulation tool so that I could do a proper review.

On Thu, Jun 10, 2021 at 9:49 PM Patrick Diehl ***@***.***>
wrote:

> Hey, @balos1 <https://github.com/balos1> @joeweaver
> <https://github.com/joeweaver> @justingray <https://github.com/justingray>
> @atismer <https://github.com/atismer> @zeroset
> <https://github.com/zeroset> would you be interested to review this paper?
>
