### 2021-06-11 #3307 @jacobmerson
@diehlpk Simmetrix should not be needed to review the paper.

If you compile `model-traits` without Simmetrix support the only thing that's missing is the Simmetrix (smd) reader. The YAML reader is most likely preferred by 99% of people since it's not linking against a proprietary library, is human readable, etc. Actually, one of my goals is that it should be easy to add new readers and writers because there is a huge variety of formats in which people store simulation model setup information.

I provide the utilities to read the propietary Simmetrix smd files because at RPI/SCOREC our simulation procedures have typically been tightly bound to using Simmetrix meshing software and model attribute libraries.

I'm happy to provide a more detailed description on how the components interact if it's helpful.
