### 2021-06-11 #3307 @joeweaver
I'm aware Simmetrex isn't needed  The issue is that while I see the
framework can be used generally, I don't have a concrete example within
which to use it.

I can see the potential of this project to define to models in a manner
independent of the simulator, but to evaluate that I'd either have to see
concrete example implementations of some readers and writers or implement
them myself (Thinking if something like generating/modifying OpenFOAM cases
from the general traits YAML). In the latter case, I unfortunately don't
have the time to do that.

It's also entirely possible that I'm misunderstanding the intended use
cases, but that'd probably be a good reason for me to not review.

None of this is a critique of the project itself.  Patrick, I like the idea
of what you're doing, a few years ago I would've benefitted greatly from a
system which could take generic definitions and produce equivalent
simulator-spevific case files. I'm just not the right person to review it.
