### 2021-06-15 #3307 @diehlpk
> Generally yes, I'm interested to review the paper. But when reading the summary of the paper I was thinking if my limited experience in FEM is sufficient.

@jacobmerson Do you think a deep experience in FEM necessary?
