### 2021-06-15 #3307 @diehlpk
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1080/10255842.2018.1538414 is OK
> - 10.1145/2814935 is OK
>
> MISSING DOIs
>
> - 10.1002/(sici)1097-0207(19970515)40:9<1573::aid-nme128>3.0.co;2-9 may be a valid DOI for title: A General Topology-Based Mesh Data Structure
> - 10.1007/s003660200030 may be a valid DOI for title: Attribute Management System for Engineering Analysis
>
> INVALID DOIs
>
> - None
> ```

@jacobmerson Please add the two missing DOIs to the paper.
