### 2021-06-16 #3307 @jacobmerson
@diehlpk A knowledge of FEM is not necessary. `model-traits` addresses simulation model setup and the association between a geometric model and the traits or attributes that define the simulation. The code is applicable to setting up any boundary value problem where the simulation setup should take place on geometric entities e.g., boundary conditions. There is no FEM specific code included in `model-traits`.

So, experience setting up CFD models, FEM models, BEM models, etc. should be adequate.
