+++ paper.bib
@@ -6,6 +6,7 @@
   volume = {40},
   pages = {1573--1596},
   publisher = {{Wiley Online Library}},
+  doi = {10.1002/(SICI)1097-0207(19970515)40:9<1573::AID-NME128>3.0.CO;2-9},
   journal = {International Journal for Numerical Methods in Engineering},
   number = {9}
 }
@@ -59,6 +60,7 @@
   volume = {18},
   pages = {339--351},
   publisher = {{Springer}},
+  doi = {10.1007/s003660200030},
   journal = {Engineering with computers},
   number = {4}
 }
