### 2021-06-16 #3307 @jacobmerson
@diehlpk I added the missing DOIs
