### 2021-06-21 #3389 @whedon
[REVIEW]: model-traits: Model attribute definitions for scientific simulations in C++
**Submitting author:** @jacobmerson (<a href="http://orcid.org/0000-0002-6813-6532">Jacob Merson</a>)
**Repository:** <a href="https://github.com/jacobmerson/model-traits" target ="_blank">https://github.com/jacobmerson/model-traits</a>
**Version:** v0.2.0
