### 2021-06-24 #3389 @diehlpk
@daniellivingston Yes, please find the links to the PDF above.
