### 2021-06-28 e391ef9
Update figures, fix Prof. Shephard initial
Figures are updated for consistency. Node names are surrounded in
quotation marks, and no node specifically has type:, or name: headings.
