### 2021-06-28 #3389 @jacobmerson
Some minor clarifying changes were made to the figures.
