### 2021-08-07 jacobmerson/model-traits#5 @pratikvn
JOSS review
I am a reviewer for the [JOSS paper](https://github.com/openjournals/joss-reviews/issues/3389) and will use this issue to track my review and provide some suggestions.

+ [x] Spack install
+ [x] Source install

### Functionality:

#### IO Backends:
+ [x] YAML
+ [ ] Simmetrix

#### Examples usage
+ [x] Basic example
+ [x] Core example
+ [ ] Example with PUMI
+ [x] Example with YAML
