### 2021-08-07 jacobmerson/model-traits#5 @pratikvn
The software and the code mostly looks good to me. This library is well-written and I believe will be useful to many.

Some issues I faced and some suggestions:

1. I am not sure how to verify the Simmetrix related functionality.
2. There is no pumi.yaml and related pumi mesh file for the pumi example. I think it would be useful to add these to the inputs folder.
3. I think it is a good idea to copy over the `inputs/` folder from the source `examples/` directory to the `build/` folder, so that it is easier to run the examples with the inputs.
4. `basic_usage.cc` and `core_usage.cc` have inline comments explaining these usage. `pumi.cc` and `yaml_with_backend.cc` lack these comments and I think it would be useful to the users to have this kind of inline comments, especially in the examples.
5. You mostly have done the heavy-lifting with respect to adding the tests and examples and so on. I think it would be very useful to add a CI step to make sure the tests and examples are running as expected. [I tried to add a basic step here and you can use that as a starting point, if you like](https://github.com/pratikvn/model-traits/runs/3270712705).
