### 2021-08-09 jacobmerson/model-traits#5 @jacobmerson
Thanks for taking a look at this and setting up the basic CI!

Unfortunately the Simmetrix libraries are proprietary. We briefly discussed the issue in the [JOSS pre-review](https://github.com/openjournals/joss-reviews/issues/3307#issuecomment-859726441) and my understanding is that the editor @diehlpk is OK with leaving the Simmetrix bits out of the review since it's not a required component. I'd like to leave the Simmetrix reader in the main model-traits repo since most of my coworkers at RPI use the Simmetrix utilities to set up their models.

I'm in the process of addressing your other concerns.
