+++ examples/yaml_with_backend.cc
@@ -8,11 +8,22 @@ int main(int argc, char *argv[]) {
   }
   std::string filename{argv[1]};
 
+  // Create a pointer to a YAML backend instance.
+  // This instance will be used to set parameters during the writing phase.
   auto yaml_backend = std::make_unique<mt::YAML>();
-  auto model_traits = mt::ReadFromFile(filename, yaml_backend.get());
 
+  // load the model traits from yaml file
+  auto model_traits = mt::ReadFromFile(filename, yaml_backend.get());
+  
+  // set the model traits prefix type which will be written to output yaml file
   yaml_backend->model_traits_prefix = "My Application";
+
+  // set the backend to write the geometry type in every model type rather than
+  // once at the beginning of the case
   yaml_backend->write_geometry_type_per_model_trait = true;
+
+  // write model traits to stdout using the yaml_backend with it's modified settings
   mt::WriteToStream(model_traits.get(), std::cout, yaml_backend.get());
+
   return 0;
 }
