### 2021-08-09 jacobmerson/model-traits#5 @jacobmerson
@pratikvn I believe that I have addressed all of the points that you have brought up.
