### 2021-08-09 jacobmerson/model-traits#5 @pratikvn
> Unfortunately the Simmetrix libraries are proprietary. We briefly discussed the issue in the JOSS pre-review and my understanding is that the editor @diehlpk is OK with leaving the Simmetrix bits out of the review since it's not a required component. I'd like to leave the Simmetrix reader in the main model-traits repo since most of my coworkers at RPI use the Simmetrix utilities to set up their models.

I see. Okay, I will then leave that out. Thanks for the information.

Thanks for addressing my concerns and adding a CI job. Everything looks good from my side and hence I will close this issue.
