### 2021-08-23 #3389 @jacobmerson
fixed a typo in the acknowledgements
