### 2021-08-23 #3389 @diehlpk
@jacobmerson I was wondering if it would help to show some sample YAML file in the paper and explain some of the features?
