+++ paper.bib
@@ -77,6 +77,7 @@
   author = {O'Leary, Patrick},
   year = {2014},
   institution = {{Kitware, Inc., Clifton Park, NY (United States)}}
+  doi = {10.2172/1349054}
 }
 
 @misc{partowExprtk2021,
