### 2021-08-23 e605906
add city, state, country to affiliation
