### 2021-08-23 #3389 @jacobmerson
@diehlpk I fixed the DOI and the affiliation. Would the following address your concerns about YAML and Simmetrix? YAML is a file format spec similar to JSON.

- Would you be happy if the sentence says "YAML file" and has a [citation to the spec](https://yaml.org/spec/1.2/spec.html)?
- For Simmetrix I discuss and cite earlier in the paper, would you be happy with just citing the Simmetrix website again?

I'm happy to add a description of a YAML "generic" input file to the paper if you think it's within scope. I do provide [example usage in the project](https://github.com/jacobmerson/model-traits/blob/main/examples/inputs). These files are fairly descriptive, but I could add more details about them in the project documentation.
