### 2021-08-23 #3389 @diehlpk
> I'm happy to add a description of a YAML "generic" input file to the paper if you think it's within scope. I do provide [example usage in the project](https://github.com/jacobmerson/model-traits/blob/main/examples/inputs). These files are fairly descriptive, but I could add more details about them in the project documentation.

I think adding some text that examples can be found here and link to the examples would help. So people can access the examples directly from the paper.
