### 2021-08-23 #3389 @diehlpk
>     * Would you be happy if the sentence says "YAML file" and has a [citation to the spec](https://yaml.org/spec/1.2/spec.html)?

My concern is that not all know what YAML is. So you might, can add some explanation that it is some file format which is text-based and easy to read?
