### 2021-08-23 #3389 @diehlpk
>     * For Simmetrix I discuss and cite earlier in the paper, would you be happy with just citing the Simmetrix website again?

@jacobmerson Maybe mention some more details about the file format? I think if the reader does not know Simmetrix, the reader might not know what kind of format it is.
