+++ paper.bib
@@ -76,7 +76,7 @@
   title = {Open-Source Integrated Design-Analysis Environment for Nuclear Energy Advanced Modeling \& Simulation Phase {{I}} Final Report},
   author = {O'Leary, Patrick},
   year = {2014},
-  institution = {{Kitware, Inc., Clifton Park, NY (United States)}}
+  institution = {{Kitware, Inc., Clifton Park, NY (United States)}},
   doi = {10.2172/1349054}
 }
 
@@ -112,5 +112,14 @@
   urldate = {2021-05-13},
   language = {en-US}
 }
+
+@misc{ben-kikiYAMLAinMarkup,
+  title = {{{YAML Ain}}'t {{Markup Language}} ({{YAML}}\texttrademark ) {{Version}} 1.2},
+  author = {{Ben-Kiki}, Oren and Evans, Clark and {d{\"o}t Net}, Ingy},
+  language = {en},
+}
+
+
+
 
 
