+++ paper.md
@@ -39,7 +39,9 @@ The first graph representation, also called the unassociated graph, is optimized
 
 ![Associated Graph representation of the "uniaxial compression" analysis case. Node types are listed on all nodes. If a node has an optional name it is surrounded in quotations.\label{fig:associated}](associated_graph.pdf)
 
-The library provides a YAML reader and writer and a smd (Simmetrix native attribute file) reader. The implementations of these readers and writer use the extensible IO interface and can be used as an example for alternative IO formats. To add an IO format the user only needs to provide template specializations for the `Read` and `Write` functions which are visible in the `mt` namespace. Additionally, `model-traits` supports the use of arbitrary geometric entity types which do not require wrapper classes or any class inheritance.
+Currently `model-traits` provides two types of readers (YAML and smd) and one type of writer (YAML), however, the library is designed to allow new IO formats to easily be added. YAML is a text based data serialization language similar to JSON which is designed to be human readable [@ben-kikiYAMLAinMarkup]. The YAML input files use keywords to define nodes in the model graph, examples of which can be found in the [`model-traits` examples](https://github.com/jacobmerson/model-traits/tree/main/examples/inputs). The the smd reader can read the proprietary attribute file format that the Simmetrix Simulation Modeling Suite™ writes [@SimmetrixIncMesh].
+
+The implementations of these readers and writer use the extensible IO interface and can be used as an example for alternative IO formats. To add an IO format the user only needs to provide template specializations for the `Read` and `Write` functions which are visible in the `mt` namespace. Additionally, `model-traits` supports the use of arbitrary geometric entity types which do not require wrapper classes or any class inheritance.
 
 During the design phase a trade off between having an open set of IO backends or an open set of model trait types and shapes had to be made. Based on the authors modeling experience, there is a finite set of model trait types and shapes that are used in practice, whereas there is an unbounded set of analysis code input file formats. Therefore, the following data shapes are supported: scalar, dynamic sized vector, and dynamic sized dense matrix. Each of these shapes can have any of the following underlying types: double or float, boolean, integer, string, null, and expressions.
 
