### 2021-08-24 #3389 @jacobmerson
@diehlpk I realized you also wanted a link to the new version with changes from review. See here https://github.com/jacobmerson/model-traits/releases/tag/v0.2.0
