+++ paper.md
@@ -18,34 +18,34 @@ bibliography: paper.bib
 ---
 
 # Summary
-`model-traits` is a C++ library for setting up scientific models and computational analysis. It provides a minimal API for applying boundary conditions (or other attributes) to the geometry of a model. `model-traits` can either be used directly as a library, or, can be used to generate input files for an existing analysis code. The library design is optimized to make adding new input and output file formats easy and maintainable without patching the core library. 
+`model-traits` is a C++ library for setting up scientific models and computational analysis. It provides a minimal API for applying boundary conditions (or other attributes) to the geometry of a model. `model-traits` can either be used directly as a library, or can be used to generate input files for an existing analysis code. The library design is optimized to make adding new input and output file formats easy and maintainable without patching the core library. 
 
 # Statement of need
 Setting up scientific simulations is often a time consuming process that typically requires hand crafting input files. Most analysis codes take a mesh-first approach to model setup, and mesh storage. When setting up a model for analysis this typically means that the user will have to recreate the input deck for any change in the mesh even if the desired boundary conditions are identical. This problem also exists with many commercial FEM codes such as Abaqus and LS-Dyna [@smithAbaqusTheoryGuide2009;@hallquistLSDYNATheoryManual2006].
 
-The essence of the problem is that the mesh is the wrong level of abstraction to apply boundary conditions. This problem also exists with respect to using a mesh-first approach to mesh databases. Many modern meshing databases such as PUMI [@ibanezPUMIParallelUnstructured2016] and MeshSim [@SimmetrixIncMesh] have shown that for many operations model geometry is a preferable level of abstraction and therefore they store the relationship between a meshes' entities and the geometric model topology [@beallGeneralTopologybasedMesh1997]. This relationship allows for much more robust mesh adaption that can correctly refine the mesh to better approximate the geometric model.
+The essence of the problem is that the mesh is the wrong level of abstraction to apply boundary conditions. This problem also exists with respect to using a mesh-first approach to mesh databases. Many modern meshing databases such as PUMI [@ibanezPUMIParallelUnstructured2016] and MeshSim [@SimmetrixIncMesh] have shown that for many operations model geometry is a preferable level of abstraction and therefore they store the relationship between a mesh's entities and the geometric model topology [@beallGeneralTopologybasedMesh1997]. This relationship allows for much more robust mesh adaption that can correctly refine the mesh to better approximate the geometric model.
 
-`model-traits` takes this geometry-first approach to applying various traits on the model such as boundary conditions, solver settings, and other data which is necessary to run an analysis. It is designed as an open source alternative to the model attribute definition provided in the Simmetrix GeomSim package [@obaraAttributeManagementSystem2002]. Kitware also develops an open source tool for model attribute definition called the Simulation Modeling Toolkit (SMTK) [@olearyOpensourceIntegratedDesignanalysis2014;@obaraComputationalModelBuilder2015]. The design of SMTK uses a deep inheritance hierarchy which makes its use challenging for users who have limited experience with the SMTK codebase. The goal for SMTK's design is also broader, encompassing GUI model setup and a general interface for 3rd party CAD kernels. The `model-traits` library is more narrowly focused on the association between a given model trait and a geometric model topological entity. Although `model-traits` does not provide an interface for any CAD kernels; CAD types, such as a pointer to a CAD kernel's face topological object, can pass through the library without modification through a templated interface. A typical use case is to use two integral types to denote a geometric entities dimension and unique ID rather than explicitly using a CAD kernel in the analysis code. 
[…]
