### 2021-08-24 #3389 @danielskatz
👋 @jacobmerson - I'm the AEiC on duty this week, and am finishing the processing of your submission. I've suggested a number of changes to the paper in https://github.com/jacobmerson/model-traits/pull/8.  Please either merge this or let me know what you disagree with, then we can complete the publication process.
